backend/requirements.txt
# Core
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
python-dotenv==1.0.0

# ML & Data Processing
numpy==1.26.3
pandas==2.1.4
scikit-learn==1.4.0
lightgbm==4.2.0
torch==2.1.2
transformers==4.36.2

# Technical Analysis
ta==0.11.0
pandas-ta==0.3.14b

# AI & NLP
google-generativeai==0.3.2
httpx==0.26.0

# Utilities
python-multipart==0.0.6
aiohttp==3.9.1
